pyproject.toml
[project]
name = "artlake"
dynamic = ["version"]
description = "ArtLake — automated art event discovery for professional artists"
requires-python = ">=3.12, <3.13"
dependencies = [
    "cffi==1.17.1",
    "cloudpickle==3.1.1",
    "numpy==2.4.0",
    "pandas==2.3.0",
    "pyarrow==22.0.0",
    "databricks-sdk==0.85.0",
    "pydantic==2.11.7",
    "loguru==0.7.3",
    "python-dotenv==1.1.1",
    "databricks-vectorsearch==0.63",
    "openai==2.8.0",
    "databricks-mcp==0.4.0",
    "backoff==2.2.1",
    "mlflow==3.8.1",
    "databricks-agents==1.8.2",
    "pyyaml==6.0.3",
    "typer==0.24.1",
    "ddgs==9.11.4",
    "requests==2.32.3",
    "beautifulsoup4==4.13.4",
    "dateparser==1.2.1",
    "geopy==2.4.1"
]

[project.scripts]
artlake = "artlake.cli:main"
artlake-generate-queries = "artlake.search.generate:main"
artlake-search = "artlake.search.web:main"
artlake-search-social = "artlake.search.social:main"
artlake-dedup = "artlake.search.dedup:main"
artlake-scrape-pages = "artlake.events.scrape:main"
artlake-extract-dates = "artlake.events.extract_dates:main"
artlake-generate-language-patterns = "artlake.events.extract_dates:main_generate_patterns"
artlake-geocode = "artlake.events.geocode:main"
artlake-categorise = "artlake.events.categorise:main"
artlake-generate-category-examples = "artlake.events.categorise:main_generate_examples"
artlake-download-artifacts = "artlake.artifacts.download:main"
artlake-process-artifacts = "artlake.artifacts.extract:main"
artlake-enrich-events = "artlake.events.enrich:main"
artlake-enrich-artifacts = "artlake.artifacts.enrich:main"
artlake-translate = "artlake.events.translate:main"

[project.optional-dependencies]
dev = ["databricks-connect>=17.0, <18",
       "ipykernel>=6.29.5, <7",
       "pip>=25.0.1, <26",
       "pre-commit>=4.1.0, <5",
       "claudetracing>=0.2.0, <1",]

ci = ["pre-commit>=4.1.0, <5",
      "pytest>=8.3.4, <9",
      "pytest-cov>=6.1.1, <7"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"artlake" = ["../../../project_config.yml"]

[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["artlake*"]

[tool.setuptools.dynamic]
version = { file = "version.txt" }

[tool.pytest.ini_options]
addopts = "-s --no-header --no-summary --cov=artlake --cov-report=term-missing --cov-fail-under=80 -m 'not integration'"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: marks tests that require a live Databricks workspace (deselect with '-m \"not integration\"')",
]

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = [
    "F",    # pyflakes rules
    "E",    # pycodestyle error rules
    "W",    # pycodestyle warning rules
    "B",    # flake8-bugbear rules
    "I",    # isort rules
    "UP",   # pyupgrade rules
    "SIM",  # flake8-simplify rules
    "ERA",  # eradicate rules
    "C",    # pylint convention rules
    "ANN"   # flake8-annotations rules
]

ignore = [
    "D107",   # Missing docstring in `__init__`
    "D203",   # incorrect-blank-line-before-class
    "D213",   # multi-line-summary-second-line
    "C901",   # too complex
    "ERA001", # commented out code
    "ANN204",
]

[tool.ruff.format]
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
packages = ["artlake"]
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = [
    "backoff",
    "databricks.sdk",
    "databricks.sdk.*",
    "bs4",
    "requests",
    "dateparser",
    "geopy",
    "geopy.*",
    "ddgs",
    "ddgs.*",
    "pandas",
    "pyspark",
    "pyspark.*",
]
ignore_missing_imports = true
